Rename 'service' notion to 'investigation'
