+++ pkg/investigations/ccam/ccam.go
