+++ pkg/investigations/ccam/ccam_suite_test.go
