+++ pkg/investigations/investigation.go
